Factor out mark-functions.h / update-functions.h
---
