+++ compiler/next/include/chpl/queries/mark-functions.h
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2021 Hewlett Packard Enterprise Development LP
+ * Other additional copyright holders may be indicated within.
+ *
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+  This file implements the generic chpl::mark.
+ */
+
+#ifndef CHPL_QUERIES_MARK_FUNCTIONS_H
+#define CHPL_QUERIES_MARK_FUNCTIONS_H
+
+#include <cassert>
+#include <cstring>
+#include <functional>
+#include <string>
+#include <unordered_map>
+#include <vector>
+#include "chpl/util/memory.h"
+
+namespace chpl {
+class Context;
+
+
+template<typename T> struct mark {
+  void operator()(Context* context, const T& keep) const { }
+};
+
+
+} // end namespace chpl
+
+#endif
